refactored hek stuff into a dedicated conversion
