+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023/__init__.py
@@ -2,5 +2,4 @@ from .behaviorinterface import BehaviorInterface
 from .behavioralsyllableinterface import BehavioralSyllableInterface
 from .fiberphotometryinterface import FiberPhotometryInterface
 from .optogeneticinterface import OptogeneticInterface
-from .hekinterface import HEKInterface
 from .nwbconverter import NWBConverter
