+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023/nwbconverter.py
@@ -4,7 +4,6 @@ from .behaviorinterface import BehaviorInterface
 from .fiberphotometryinterface import FiberPhotometryInterface
 from .optogeneticinterface import OptogeneticInterface
 from .behavioralsyllableinterface import BehavioralSyllableInterface
-from .hekinterface import HEKInterface
 
 
 class NWBConverter(NWBConverter):
@@ -15,5 +14,4 @@ class NWBConverter(NWBConverter):
         BehavioralSyllable=BehavioralSyllableInterface,
         FiberPhotometry=FiberPhotometryInterface,
         Optogenetic=OptogeneticInterface,
-        HEK=HEKInterface,
     )
