+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023_HEK/convert_session.py
@@ -2,10 +2,10 @@
 from pathlib import Path
 from typing import Union
 from neuroconv.utils import dict_deep_update, load_dict_from_file
-from datta_lab_to_nwb import markowitz_gillis_nature_2023
+from datta_lab_to_nwb import markowitz_gillis_nature_2023_HEK
 import shutil
 from pynwb import NWBHDF5IO
-from datta_lab_to_nwb.markowitz_gillis_nature_2023.postconversion import reproduce_figS1abcd
+from datta_lab_to_nwb.markowitz_gillis_nature_2023_HEK.postconversion import reproduce_figS1abcd
 
 
 def session_to_nwb(
@@ -28,11 +28,11 @@ def session_to_nwb(
     conversion_options = {
         "HEK": dict(),
     }
-    converter = markowitz_gillis_nature_2023.NWBConverter(source_data=source_data)
+    converter = markowitz_gillis_nature_2023_HEK.NWBConverter(source_data=source_data)
     metadata = converter.get_metadata()
 
     # Update metadata
-    paper_metadata_path = Path(__file__).parent / "markowitz_gillis_nature_2023_metadata.yaml"
+    paper_metadata_path = Path(__file__).parent / "markowitz_gillis_nature_2023_HEK_metadata.yaml"
     paper_metadata = load_dict_from_file(paper_metadata_path)
     metadata = dict_deep_update(metadata, paper_metadata)
