+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023_HEK/markowitz_gillis_nature_2023_HEK_metadata.yaml
@@ -1,0 +1,34 @@
+NWBFile:
+  related_publications:
+    - https://doi.org/10.1038/s41586-022-05611-2
+  experiment_description: Abstract -- Spontaneous animal behaviour is built from action modules that are concatenated by the brain into sequences1,2. However, the neural mechanisms that guide the composition of naturalistic, self-motivated behaviour remain unknown. Here we show that dopamine systematically fluctuates in the dorsolateral striatum (DLS) as mice spontaneously express sub-second behavioural modules, despite the absence of task structure, sensory cues or exogenous reward. Photometric recordings and calibrated closed-loop optogenetic manipulations during open field behaviour demonstrate that DLS dopamine fluctuations increase sequence variation over seconds, reinforce the use of associated behavioural modules over minutes, and modulate the vigour with which modules are expressed, without directly influencing movement initiation or moment-to-moment kinematics. Although the reinforcing effects of optogenetic DLS dopamine manipulations vary across behavioural modules and individual mice, these differences are well predicted by observed variation in the relationships between endogenous dopamine and module use. Consistent with the possibility that DLS dopamine fluctuations act as a teaching signal, mice build sequences during exploration as if to maximize dopamine. Together, these findings suggest a model in which the same circuits and computations that govern action choices in structured tasks have a key role in sculpting the content of unconstrained, high-dimensional, spontaneous behaviour.
+  keywords:
+    - Basal Ganglia
+    - Neural circuits
+    - Reward
+  institution: Harvard Medical School
+  lab: Datta
+  experimenter:
+    - Markowitz, Jeffrey E.
+    - Gillis, Winthrop F.
+    - Jay, Maya
+    - Wood, Jeffrey
+    - Harris, Ryley W.
+    - Cieszkowski, Robert
+    - Scott, Rebecca
+    - Brann, David
+    - Koveal, Dorothy
+    - Kula, Tomasz
+    - Weinreb, Caleb
+    - Osman, Mohammed Abdal Monium
+    - Pinto, Sandra Romero
+    - Uchida, Naoshige
+    - Linderman, Scott W.
+    - Sabatini, Bernardo L.
+    - Datta, Sandeep Robert
+Subject:
+  species: Homo sapiens
+  age: fetus
+  sex: F
+  subject_id: HEK293
+  description: HEK 293 cells (ATCC, cells were validated by ATCC via short tandem repeat analysis and were not tested for mycoplasma) were transfected with the dLight1.1 plasmid (Addgene 111067-AAV5) using Mirus TransIT-LT1 (MIR 2304).
