+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023_HEK/nwbconverter.py
@@ -1,0 +1,11 @@
+"""Primary NWBConverter class for this dataset."""
+from neuroconv import NWBConverter
+from .hekinterface import HEKInterface
+
+
+class NWBConverter(NWBConverter):
+    """Primary conversion class."""
+
+    data_interface_classes = dict(
+        HEK=HEKInterface,
+    )
