+++ src/datta_lab_to_nwb/markowitz_gillis_nature_2023_HEK/postconversion/reproduce_figS1abcd.py
